♻️ refactor(new.tsx): split components into separate files for better maintainability
The commit reorganizes the large new.tsx file by extracting several components into their own files:
- Move PhotoGrid to its own component
- Create EventPreview component for the preview section
- Create NewEventHeader component for the header section
- Create CaptureEventButton component for the action button
- Improve styling consistency using cn utility
- Remove unused imports and clean up code structure
